% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractModelReact.R
\name{extractModelReact}
\alias{extractModelReact}
\title{extract reactions from a sybil model as a list of strings}
\usage{
extractModelReact(object)
}
\arguments{
\item{model}{A metabolic model in the form of "modelorg" class from sybil package}
}
\description{
This function is based in printReaction function from sybil, it
extracts all reactions from a sybil model as a list of strings with the fromat:
\code{"H2O[c] + Urea-1-carboxylate[c] <=> 2 CO2[c] + 2 NH3[c]"}

Where arrows and plus signs are surrounded by a "space character".
It is also expected that stoichiometry coefficients are surrounded by spaces, (nothe the "2" before the CO2[c] or the NH3[c]).
It also expects arrows to be in the form "\code{=>}" or "\code{<=>}".
Meaning that arrows like "\code{==>}", "\code{<==>}", "\code{-->}" or "\code{->}" will not be parsed and will lead to errors.
}
\examples{
\dontrun{
# Get the sybil model
library("sybil")
data("Ec_core")

# Extract reactions
react <- extractModelReact(Ec_core)
}
}
\seealso{
\code{additionCost} function documentation.
}
\author{
Nicolas Mendoza Mejia <nimendozam@unal.edu.co>
}
