% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gapFill.R
\name{gapFill}
\alias{gapFill}
\title{Find and fill gaps in a metabolic network}
\usage{
gapFill(reactionList, reference, limit = 0.25, woCompartment = FALSE,
  consensus = FALSE)
}
\description{
This function identifies the gaps and fills it from the stoichiometric reactions of a reference metabolic reconstruction using a weighting function
}
\author{
Kelly Botero <kjboteroo@unal.edu.co> - Maintainer: Daniel Osorio <dcosorioh@unal.edu.co>
}

