% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gapFind.R
\name{gapFind}
\alias{gapFind}
\title{Find gaps in a metabolic network}
\usage{
gapFind(reactionList, removeExchange = FALSE)
}
\description{
This function identifies the gaps (metabolites not produced or not consumed in any other reaction or just involved in one reaction) for a set of stoichometric reactions
}
\author{
Daniel Osorio <dcosorioh@unal.edu.co>
}
