% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additionCost.R
\name{additionCost}
\alias{additionCost}
\title{Calculate the addition cost of a stoichiometric reaction in a metabolic network}
\usage{
additionCost(reaction, reference)
}
\author{
Daniel Camilo Osorio <dcosorioh@unal.edu.co>
}

