% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getReference.R
\name{getReference}
\alias{getReference}
\title{Download all the set of gene-associated stoichiometric reactions for a specific organism from the KEGG database}
\usage{
getReference(organism, sep = " ; ")
}
\author{
Kelly Johana Botero <kjboteroo@unal.edu.co> - Mantainer: Daniel Camilo Osorio <dcosorioh@unal.edu.co>
}

